+++ components/Button/__tests__/index.test.jsx
@@ -1,0 +1,11 @@
+import React from 'react';
+import { render } from 'enzyme';
+import toJson from 'enzyme-to-json';
+import Button from '../index';
+
+describe('Button', () => {
+  it('renders correctly', () => {
+    const wrapper = render(<Button>foo</Button>);
+    expect(toJson(wrapper)).toMatchSnapshot();
+  });
+});