+++ components/Panel/__tests__/index.test.jsx
@@ -1,0 +1,18 @@
+import React from 'react';
+import { render } from 'enzyme';
+import toJson from 'enzyme-to-json';
+import Panel from '../index';
+
+describe('Panel', () => {
+  it('renders correctly', () => {
+    const wrapper = render(
+      <Panel>
+        <Panel.Header>
+          <Panel.Title>title</Panel.Title>
+        </Panel.Header>
+        <Panel.Body>body</Panel.Body>
+      </Panel>
+    );
+    expect(toJson(wrapper)).toMatchSnapshot();
+  });
+});