+++ components/index.js
@@ -11,6 +11,7 @@ export { default as Loading } from './Loading';
 export { default as Message } from './Message';
 export { default as Mask } from './Mask';
 export { default as Modal } from './Modal';
+export { default as NoticeBar } from './NoticeBar';
 export { default as Popup } from './Popup';
 export { default as Progress } from './Progress';
 export { default as Panel } from './Panel';